Made the way the decorator pattern is implemented more clear. <layout
