--- src/main/java/nl/joshuaslik/tudelft/SEM/control/gameObjects/pickup/powerup/IModifier.java
@@ -1,12 +1,0 @@
-///*
-// * To change this license header, choose License Headers in Project Properties.
-// * To change this template file, choose Tools | Templates
-// * and open the template in the editor.
-// */
-//package nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.powerup;
-//
-///**
-// * @author faris
-// */
-//public interface IModifier {
-//}
