+++ src/main/java/nl/joshuaslik/tudelft/SEM/control/gameObjects/pickup/powerup/bubble/bubbleMods/BubbleSlowdown.java
@@ -13,7 +13,7 @@ import nl.joshuaslik.tudelft.SEM.control.gameObjects.pickup.powerup.bubble.Abstr
 public class BubbleSlowdown extends AbstractBubbleDecorator {
 	
 	@Override
-	public double getBubbleSpeedModifier() {
-		return 0.5 * getChild().getBubbleSpeedModifier();
+	public double getBubbleSpeedMultiplier() {
+		return 0.5 * getChild().getBubbleSpeedMultiplier();
 	}
 }
